feat: organize security views
