+++ frontend/app/pages/dashboard/settings/_views/_components/SessionCard.vue
